backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
alembic==1.13.1
pydantic[email]>=2.6
google-cloud-secret-manager
pg8000
cloud-sql-python-connector
